+++ sdk/networkcloud/arm-networkcloud/src/operations/kubernetesClusters.ts
@@ -790,16 +790,16 @@ const restartNodeOperationSpec: coreClient.OperationSpec = {
   httpMethod: "POST",
   responses: {
     200: {
-      headersMapper: Mappers.KubernetesClustersRestartNodeHeaders
+      bodyMapper: Mappers.OperationStatusResult
     },
     201: {
-      headersMapper: Mappers.KubernetesClustersRestartNodeHeaders
+      bodyMapper: Mappers.OperationStatusResult
     },
     202: {
-      headersMapper: Mappers.KubernetesClustersRestartNodeHeaders
+      bodyMapper: Mappers.OperationStatusResult
     },
     204: {
-      headersMapper: Mappers.KubernetesClustersRestartNodeHeaders
+      bodyMapper: Mappers.OperationStatusResult
     },
     default: {
       bodyMapper: Mappers.ErrorResponse